fix: login page loading state
