+++ src/components/templates/descopeMiddleware.tsx
@@ -1,6 +1,6 @@
-import React, { ReactNode, useCallback, useEffect, useState } from "react";
+import React, { ReactNode, Suspense, lazy, useCallback, useEffect, useState } from "react";
 
-import { Descope, useDescope } from "@descope/react-sdk";
+import { useDescope } from "@descope/react-sdk";
 import axios from "axios";
 import Cookies from "js-cookie";
 import psl from "psl";
@@ -13,10 +13,9 @@ import { useUserStore } from "@store/useUserStore";
 
 import { useToastStore } from "@store";
 
-import { AHref, IconSvg } from "@components/atoms";
+import { Loader } from "@components/atoms";
 
-import { AKRoundLogo } from "@assets/image";
-import { InJustTitle, LoginLogos } from "@assets/image/pages/login";
+const LoginPage = lazy(() => import("../pages/login"));
 
 export const DescopeMiddleware = ({ children }: { children: ReactNode }) => {
 	const { getLoggedInUser, setLogoutFunction } = useUserStore();
@@ -32,14 +31,11 @@ export const DescopeMiddleware = ({ children }: { children: ReactNode }) => {
 				message: t("errors.logoutFailed", { error: rootDomain.error.message }),
 				type: "error",
 			});
-
 			LoggerService.error(namespaces.ui.loginPage, t("errors.logoutFailed", { error: rootDomain.error.message }));
 
 			return;
 		}
-
 		Cookies.remove(isLoggedInCookie, { domain: getCookieDomain(rootDomain) });
-
 		window.localStorage.clear();
 		window.location.reload();
 		// eslint-disable-next-line react-hooks/exhaustive-deps
@@ -55,7 +51,6 @@ export const DescopeMiddleware = ({ children }: { children: ReactNode }) => {
 		async (event: CustomEvent<any>) => {
 			try {
 				const apiBaseUrl = getApiBaseUrl();
-
 				await axios.get(`${apiBaseUrl}/auth/descope/login?jwt=${event.detail.sessionJwt}`, {
 					withCredentials: true,
 				});
@@ -81,77 +76,14 @@ export const DescopeMiddleware = ({ children }: { children: ReactNode }) => {
 	);
 
 	const isLoggedIn = Cookies.get(isLoggedInCookie);
+
 	if (authBearer || isLoggedIn) {
 		return children;
 	}
 
-	const benefitsList = Object.values(t("rightSide.benefitsList", { returnObjects: true }));
-
 	return (
-		<div className="flex h-screen">
-			<div className="relative flex w-2/3 items-center justify-center rounded-3xl pb-32 pl-16 text-black">
-				<div className="z-10 p-5">
-					<AHref
-						ariaLabel={t("leftSide.logoText")}
-						className="absolute left-4 top-4 flex h-auto items-center"
-						href="https://autokitteh.com/"
-						relationship="noreferrer"
-						target="_blank"
-						title={t("leftSide.logoText")}
-					>
-						<IconSvg className="top-8 size-10" src={AKRoundLogo} />{" "}
-						<div className="ml-3 font-averta text-2xl font-bold">{t("leftSide.logoText")}</div>
-					</AHref>
-					<h2 className="mt-16 font-averta text-4xl font-bold">{t("rightSide.titleFirstLine")}</h2>
-
-					<div className="flex">
-						<IconSvg className="ml-4 mr-2 h-10 w-24" size="3xl" src={InJustTitle} />
-
-						<h2 className="font-averta text-4xl font-bold">{t("rightSide.titleSecondLine")}</h2>
-					</div>
-
-					<h3 className="mt-12 max-w-485 font-averta text-2xl font-bold">
-						{t("rightSide.descriptionFirstLine")}
-					</h3>
-
-					<ul className="ml-4 mt-10 font-averta text-xl font-semibold">
-						{benefitsList?.length
-							? Object.values(benefitsList).map((benefit) => (
-									<li className="mt-2 before:size-1 before:bg-black" key={benefit}>
-										{benefit}
-									</li>
-								))
-							: null}
-					</ul>
-				</div>
-				<IconSvg
-					alt="autokitteh logo with integrations"
-					className="absolute bottom-0 right-8 h-36 w-9/12 object-contain object-bottom"
-					src={LoginLogos}
-				/>
-			</div>
-
-			<div className="z-10 flex w-7/12 shrink-0 flex-col items-center justify-center rounded-l-2xl bg-gray-1250 p-8 font-averta text-white">
-				<h1 className="text-center font-averta text-4xl font-semibold">
-					{t("leftSide.welcomeTitle")}
-					<br />
-					{t("leftSide.autokittehGreenTitle")}
-				</h1>
-
-				<Descope flowId="sign-up-or-in" key={descopeRenderKey} onSuccess={handleSuccess} />
-
-				<div>
-					{t("leftSide.signupText")}{" "}
-					<AHref
-						className="font-averta text-green-800 hover:text-gray-500"
-						href="https://autokitteh.com/get-a-demo/"
-						relationship="noreferrer"
-						target="_blank"
-					>
-						{t("leftSide.signupLink")}
-					</AHref>
-				</div>
-			</div>
-		</div>
+		<Suspense fallback={<Loader isCenter />}>
+			<LoginPage descopeRenderKey={descopeRenderKey} handleSuccess={handleSuccess} t={t} />
+		</Suspense>
 	);
 };
