+++ src/interfaces/components/index.ts
@@ -44,3 +44,4 @@ export type { CheckboxProps } from "@interfaces/components/checkbox.interface";
 export type { AccordionProps } from "@interfaces/components/accordion.interface";
 export type { BadgeProps } from "@interfaces/components/badge.interface";
 export type { HrefProps } from "@interfaces/components/href.interface";
+export type { LoginPageProps } from "@interfaces/components/loginPage.interface";
