+++ src/interfaces/components/loginPage.interface.ts
@@ -1,0 +1,5 @@
+export interface LoginPageProps {
+	descopeRenderKey: number;
+	handleSuccess: (event: CustomEvent<any>) => Promise<void>;
+	t: (key: string, options?: any) => string;
+}
