+++ tsconfig.json
@@ -48,6 +48,8 @@
 			"@interfaces": ["./src/interfaces/index"],
 			"@type/*": ["./src/types/*"],
 			"@type": ["./src/types/index"],
+			"@pages/*": ["./pages/*"],
+			"@pages": ["./pages/index"],
 			"@api": ["./src/api/index"],
 			"@store/*": ["./src/store/*"],
 			"@store": ["./src/store/index"],
